Fix minor issues to pass check
Halfway through the body of the format_timedata function there is an
object that should be recognized as a data.table but instead is taken
for a data.frame only. An extra line has been added to force the code to
pick the object as a data.table.

A vignette was failing to build due to a required package -data.table-
not being loaded, and this has now been fixed.